dom/html/test/test_non-ascii-cookie.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=784367
-->
<head>
  <meta charset="utf-8">
  <title>Test for non-ASCII cookie values</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=784367">Mozilla Bug 784367</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for non-ASCII cookie values **/

SimpleTest.waitForExplicitFinish();

var gScript = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL("file_cookiemanager.js"));

function getCookieFromManager() {
  return new Promise(resolve => {
    gScript.addMessageListener("getCookieFromManager:return", function gcfm({ cookie }) {
      gScript.removeMessageListener("getCookieFromManager:return", gcfm);
      resolve(cookie);
    });
    gScript.sendAsyncMessage("getCookieFromManager", { host: location.hostname, path: location.pathname });
  });
}

var c = document.cookie;
is(document.cookie, 'abc=012©ABC\ufffdDEF', "document.cookie should be decoded as UTF-8");

var newCookie;

getCookieFromManager().then((cookie) => {
  is(cookie, document.cookie, "nsICookieManager should be consistent with document.cookie");
  newCookie = 'def=∼≩≭≧∯≳≲≣∽≸≸∺≸∠≯≮≥≲≲≯≲∽≡≬≥≲≴∨∱∩∾';
  document.cookie = newCookie;
  is(document.cookie, c + '; ' + newCookie, "document.cookie should be encoded as UTF-8");

  return getCookieFromManager();
}).then((cookie) => {
  is(cookie, document.cookie, "nsICookieManager should be consistent with document.cookie");
  var date1 = new Date();
  date1.setTime(0);
  document.cookie = newCookie + 'def=;expires=' + date1.toGMTString();
  gScript.destroy();
  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
